requirements.txt
awsebcli==3.25.3
bcrypt==5.0.0
blessed==1.24.0
botocore==1.40.75
cement==2.10.14
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
Deprecated==1.3.1
fabric==3.2.2
idna==3.11
invoke==2.2.1
jmespath==1.0.1
packaging==24.2
paramiko==4.0.0
pathspec==0.12.1
pycparser==2.23
PyNaCl==1.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
termcolor==2.5.0
urllib3==1.26.20
wcwidth==0.2.14
wrapt==2.0.1
google-genai

